WIP: Switch to glib with VariantIter Support
To pull from a repo we need to fill out a Variant object.
With the official crates, only basic and dict is implemented.
Newer versions also support the array of strings which
we do need to describe our revision to fetch.

Unfortunately moving up with gio and glib
introduces some issues with the Cancellable object.

Signed-off-by: Ralf Anton Beier <ralf_beier@me.com>